Removes sentence about language support
Exciting news! Netlify now supports serverless functions in TypeScript in addition to JavaScript and Go.

This copy edit suggests the removal of a sentence on the functions.netlify.com homepage that describes language support. The intention of removing this sentence is to ensure we won't have to continually update it or expand on it here in the future.